// frontend/src/types/errors.ts
export enum BillingErrors {
  alreadySubscribed = "alreadySubscribed",
  forbidden = "forbidden",
  userCustomerIdNotExists = "userCustomerIdNotExists",
  priceIdNotExists = "priceIdNotExists",
  customerDeleted = "customerDeleted",
  webhookSignInvalid = "webhookSignInvalid",
  errorUserNotSubscribed = "errorUserNotSubscribed",
  companyNotFound = "companyNotFound",
  companySubscriptionNotFound = "companySubscriptionNotFound",
  errorUpdateStripeSubscription = "errorUpdateStripeSubscription",
  errorRetrieveStripeSubscription = "errorRetrieveStripeSubscription",
  errorSubscriptionNotCanceled = "errorSubscriptionNotCanceled",
  errorCreateStripeCustomer = "errorCreateStripeCustomer",
  errorDeleteStripeSubscription = "errorDeleteStripeSubscription",
  customerNotExists = "customerNotExists",
  paymentMethodNotExists = "paymentMethodNotExists",
  subscriptionInProgress = "subscriptionInProgress",
  stripeUserNotExists = "stripeUserNotExists",
  subscriptionNotFound = "subscriptionNotFound",
  BillingUnhandledError = "BillingUnhandledError",
}

export enum ClientErrors {
  CompanyNotFound = "CompanyNotFound",
  Forbidden = "Forbidden",
  SubscriptionExpired = "SubscriptionExpired",
  CompanyRemoved = "CompanyRemoved",
  ClientDuplicated = "ClientDuplicated",
  CsvInvalid = "CsvInvalid",
  ImportLimitExceeded = "ImportLimitExceeded",
  FieldFromRequired = "FieldFromRequired",
  CompanyEmailInvalid = "CompanyEmailInvalid",
  InvitationsNotSent = "InvitationsNotSent",
  ClientUnhandledError = "ClientUnhandledError",
  EmailNotSent = "EmailNotSent",
  CreateInviteEmailFailed = "CreateInviteEmailFailed",
}

export enum CompanyErrors {
  LogoUploadError = "LogoUploadError",
  UserAlreadyHasCompany = "UserAlreadyHasCompany",
  CompanyNotFound = "CompanyNotFound",
  Forbidden = "Forbidden",
  CompanyUnhandledError = "CompanyUnhandledError",
  BadInvitationText = "BadInvitationText",
  BadReminderText = "BadReminderText",
  UrlInInvitationText = "UrlInInvitationText",
  UrlInReminderText = "UrlInReminderText",
  BccDelayLimitExceeded = "BccDelayLimitExceeded",
  BccEmailInUse = "BccEmailInUse",
  FingerprintRequired = "FingerprintRequired",
  CompanyVisitRestricted = "CompanyVisitRestricted",
  ReminderDelayLimitExceeded = "ReminderDelayLimitExceeded",
  ReminderDelayLessThenMinimum = "ReminderDelayLessThenMinimum",
  TitleIsProfane = "TitleIsProfane",
  BodyIsProfane = "BodyIsProfane",
  CompanyFingerprintMarkerUndefined = "CompanyFingerprintMarkerUndefined",
  CompanyIsNotTest = "CompanyIsNotTest",
}

export enum PageErrors {
  Forbidden = "forbidden",
  PageWitchNameAlreadyExists = "pageWitchNameAlreadyExists",
  PageNotFound = "pageNotFound",
  PageUnhandledError = "PageUnhandledError",
}

export enum ReviewErrors {
  ReviewNotFound = "ReviewNotFound",
  ReportSubmissionLimit = "ReportSubmissionLimit",
  Forbidden = "Forbidden",
  MessageIsProfane = "MessageIsProfane",
  InvalidToken = "InvalidToken",
  ClientFieldsRequired = "ClientFieldsRequired",
  ErrorLastReviewTimeLimit = "ErrorLastReviewTimeLimit",
  DoublicateReviewMessage = "DoublicateReviewMessage",
  CompanyNotFound = "CompanyNotFound",
  CompanySubscriptionExpired = "CompanySubscriptionExpired",
  CompanyRemoved = "CompanyRemoved",
  ReviewUnhandledError = "ReviewUnhandledError",
  ReviewEditTimeExpired = "ReviewEditTimeExpired",
}

export enum UserErrors {
  Forbidden = "forbidden",
  UserNotFound = "userNotFound",
  UserAlreadyExists = "userAlreadyExists",
  EmailNotVerified = "emailNotVerified",
  EmailAlreadyVerified = "emailAlreadyVerified",
  InvalidPassword = "invalidPassword",
  InvalidAccessToken = "invalidAccessToken",
  InvalidVerifyEmailCode = "invalidVerifyEmailCode",
  InvalidChangePasswordCode = "invalidChangePasswordCode",
  UserBanned = "userBanned",
  InvalidOldPassword = "invalidOldPassword",
  EqualPasswords = "equalPasswords",
  CompanyNotFound = "CompanyNotFound",
  BlankRequest = "BlankRequest",
  EmailNotSent = "EmailNotSent",
  InvalidEmailVerifyCode = "invalidEmailVerifyCode",
  MissingAuthorizationToken = "MissingAuthorizationToken",
  InvalidAuthorizationToken = "InvalidAuthorizationToken",
  ContactFormSendError = "ContactFormSendError",
  UserUnhandledError = "UserUnhandledError",
}

export enum WidgetErrors {
  CompanyNotFound = "CompanyNotFound",
  CompanyRemoved = "CompanyRemoved",
  SubscriptionExpired = "SubscriptionExpired",
}